requirements.txt
cachetools==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
google-api-core==2.19.1
google-api-python-client==2.137.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.63.2
httplib2==0.22.0
idna==3.7
numpy==2.0.0
pillow==10.4.0
proto-plus==1.24.0
protobuf==5.27.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pyboy==2.2.1
pyparsing==3.1.2
PySDL2==0.9.16
pysdl2-dll==2.30.2
requests==2.32.3
rsa==4.9
uritemplate==4.1.1
urllib3==2.2.2
websockets==12.0
python-dotenv==1.0.1
boto3==1.34.144
